birdflu-ensemble/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birdflu-ensemble"
version = "0.1.0"
description = "Defense-in-depth classification system for bird flu papers"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "xgboost>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.19.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "snorkel>=0.9.7",
    "loguru>=0.7.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict_optional = true